requirements.txt
flask
psycopg2-binary
flask-sqlalchemy
python-dotenv
openai
aioflask
gtts
python-telegram-bot
